+++ HISTORY.rst
@@ -7,6 +7,7 @@ History
 Unreleased
 -----------
 
+* Feature: add support for horizontal forms
 * BREAKING CHANGE: HTML in labels is now being escaped, this is consistent with django.
 * BREAKING CHANGE: EmailField and EmailInput have been removed. Use django.forms.EmailField instead.
 * Feature: add support for forms.MultipleChoiceField with CheckboxSelectMultiple as widget.